Split Gav and Ga from ArtifactCoords
